release 3.0.705

+++ TencentCloud/Lighthouse/V20200324/Models/BlueprintPrice.cs
@@ -40,7 +40,7 @@ namespace TencentCloud.Lighthouse.V20200324.Models
         /// 折扣。
         /// </summary>
         [JsonProperty("Discount")]
-        public long? Discount{ get; set; }
+        public float? Discount{ get; set; }
 
         /// <summary>
         /// 镜像折扣后总价。单位元。
