+++ TencentCloud/Lighthouse/V20200324/Models/DiscountDetail.cs
@@ -52,7 +52,7 @@ namespace TencentCloud.Lighthouse.V20200324.Models
         /// 折扣。
         /// </summary>
         [JsonProperty("Discount")]
-        public long? Discount{ get; set; }
+        public float? Discount{ get; set; }
 
         /// <summary>
         /// 具体折扣详情。
